Start rework of motion ID system to 'timestamp offset' system
Structure refactor and naming changes, small prep work.

+++ firmware/src/drivers/user_interface_types.h
@@ -93,7 +93,7 @@ typedef struct
     //information about the current move being executed
     uint8_t  profile_type;
     uint8_t  move_progress;
-    uint16_t movement_identifier;
+    uint32_t movement_identifier;
 } MotionData_t;
 
 typedef struct
